push to linux

+++ libprintf.h
@@ -11,5 +11,5 @@ int	ft_printf(const char *format, ...);
 int	ft_putnbr(int n);
 int	ft_putstr(char *s);
 int	ft_putchar(char c);
-int	ft_puthex(unsigned int  n, char c);
+int	ft_puthex(unsigned long long n, char c);
 #endif
